+++ icat/query.py
@@ -485,12 +485,13 @@ class Query(object):
         """Return a formal representation of the query.
         """
         return ("%s(%s, %s, attributes=%s, aggregate=%s, order=%s, "
-                "conditions=%s, includes=%s, limit=%s)"
+                "conditions=%s, includes=%s, limit=%s, join_specs=%s)"
                 % (self.__class__.__name__,
                    repr(self.client), repr(self.entity.BeanName),
                    repr(self.attributes), repr(self.aggregate),
                    repr(self.order), repr(self.conditions),
-                   repr(self.includes), repr(self.limit)))
+                   repr(self.includes), repr(self.limit),
+                   repr(self.join_specs)))
 
     def __str__(self):
         """Return a string representation of the query.
@@ -574,6 +575,7 @@ class Query(object):
             q.conditions[k] = self.conditions[k].copy()
         q.includes = self.includes.copy()
         q.limit = self.limit
+        q.join_specs = self.join_specs.copy()
         return q
 
     def setAttribute(self, attribute):